requirements.txt
streamlit
google-generative-ai
PyPDF2
python-dotenv
IPython
